Refactor: 클라이밍 채널 참여 <-> 취소 API 위치 수정
